Fix the build on a toolchain older than the author's
CI on a clean runner failed where the local machine passed. Three test
literals relied on Int-to-UInt64 inference that a newer Swift accepts and
the runner's does not, which is exactly the class of thing that is
invisible until someone else tries to build your project — the first
person to clone this would have hit it.

Also decodes the executable path using the length proc_pidpath reports
rather than scanning for a null terminator, which drops a deprecation
warning that had been repeating through every build.

+++ Tests/BoostKitTests/HistoryTests.swift
@@ -109,7 +109,7 @@ struct AppGrowthTests {
     @Test("Leaves a large but stable app alone")
     func ignoresLargeButSteady() {
         var g = AppGrowth()
-        for _ in 0..<40 { g.record([(id: "browser", bytes: 4096 * 1_048_576)]) }
+        for _ in 0..<40 { g.record([(id: "browser", bytes: UInt64(4096) * 1_048_576)]) }
         #expect(g.growth(of: "browser") == nil)
     }
 
@@ -132,9 +132,9 @@ struct AppGrowthTests {
     @Test("An app that exits stops being tracked")
     func forgetsDeadApps() {
         var g = AppGrowth()
-        for _ in 0..<10 { g.record([(id: "gone", bytes: 500 * 1_048_576)]) }
+        for _ in 0..<10 { g.record([(id: "gone", bytes: UInt64(500) * 1_048_576)]) }
         #expect(g.samples(for: "gone") == 10)
-        g.record([(id: "other", bytes: 100 * 1_048_576)])
+        g.record([(id: "other", bytes: UInt64(100) * 1_048_576)])
         #expect(g.samples(for: "gone") == 0)
     }
